requirements.txt
# Damn Vulnerable Startup (DVS) - Dependencies
# Core framework
fastapi==0.115.14
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# Database
sqlalchemy==2.0.48

# Templates
jinja2>=3.1.0

# Auth and crypto
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4

# Data files
pyyaml>=6.0

# Environment config
pydantic-settings>=2.0.0
itsdangerous>=2.0.0

# Syntax highlighting (View Source feature)
pygments>=2.18.0

# HTTP client (SSRF challenges)
httpx>=0.27.0

# HTML sanitization (tech_lead tier XSS defense)
bleach>=6.1.0

# LLM providers (optional, mock is default)
openai>=1.50.0
anthropic>=0.39.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Linting
ruff>=0.8.0
